fix: rename `queryKey`, move withdraw modal into its own file
